Adding DI and forms
