+++ MineSweeperClientBlazor/Components/MineCounter.razor
@@ -3,15 +3,17 @@
 <style>
     .AmountOfMines
     {
-        @* position: relative;
-        font-family: monospace;
-        background-color: black;
-        text-decoration: none;
-        color: green;
-        font-size: 1.5em; *@
         padding: 20px;
-        margin-left: 15px;
+        margin-left: 2px;
         width: 80px;
+        height: 80px;
+        border: solid 1px gray
+    }
+    .AmountOfMinesTime
+    {
+        padding: 20px;
+        margin-left: 2px;
+        width: 160px;
         height: 80px;
         border: solid 1px gray
     }
@@ -27,7 +29,7 @@
 
 </style>
 <div class="AmountOfMines ">
-    <div class="AmountOfMinesNumber blink">@AmountOfMinesLeft</div>
+    <div class="AmountOfMinesNumber">@AmountOfMinesLeft</div>
 </div>
