+++ MineSweeperClientBlazor/Pages/AmountOfMines.razor
@@ -1,4 +1,31 @@
 @page "/amountOfMines"
+
+@using MineSweeperLogic
+
+@inject IGame _game
+@inject NavigationManager NavManager
+
     <div class="container text-center">
         <h1>Set the amount of mines</h1>
+        <EditForm Model="@_game" OnValidSubmit="@HandleValidSubmit">
+        <DataAnnotationsValidator />
+        <ValidationSummary />
+        <p>
+                <label>
+                    Amount:
+                    <InputNumber @bind-Value="_game.AmountOfMines" />
+                </label>
+            </p>
+
+        <button type="submit">Submit</button>
+    </EditForm>
     </div>
+
+@code {
+
+    private void HandleValidSubmit()
+    {
+        _game.Start();
+        NavManager.NavigateTo("/minesweeper");
+    }
+}
