+++ MineSweeperClientBlazor/Pages/HeightAndWithdt.razor
@@ -1,4 +1,37 @@
 @page "/heightandwidth"
+
+@using MineSweeperLogic
+
+@inject IGame _game
+@inject NavigationManager NavManager
+
 <div class="container text-center">
     <h1>Set the height and width</h1>
-</div>
+    <EditForm Model="@_game" OnValidSubmit="@HandleValidSubmit">
+        <DataAnnotationsValidator />
+        <ValidationSummary />
+        <p>
+                <label>
+                    Height:
+                    <InputNumber @bind-Value="_game.Height" />
+                </label>
+            </p>
+            <p>
+                <label>
+                    Width:
+                    <InputNumber @bind-Value="_game.Width" />
+                </label>
+            </p>
+
+        <button type="submit">Submit</button>
+    </EditForm>
+</div> 
+
+@code {
+
+    private void HandleValidSubmit()
+    {
+        _game.Start();
+        NavManager.NavigateTo("/minesweeper");
+    }
+}
