+++ MineSweeperClientBlazor/Shared/MainLayout.razor
@@ -8,7 +8,7 @@
         </div>
             <ul class="nav justify-content-center nav-pills">
             <li class="nav-item">
-            <a class="nav-link active" href="fetchdata">MineSweeper</a>
+            <a class="nav-link active" href="minesweeper">MineSweeper</a>
             </li>
             <li class="nav-item">
                 <a class="nav-link" href="heightandwidth">Height & Width</a>
